Changed fontsize was working
