+++ timeline/src/routes/login/+page.svelte
@@ -16,90 +16,12 @@
     console.log("i forgor 💀");
 	// TODO: add forgot password functionality
   };
-
-  function IncFontSize() {
-		var elmntsToChg = document.querySelectorAll('[id=text]');
-		
-		for (let i = 0; i < elmntsToChg.length; i ++) {
-			// @ts-ignore
-			if (!elmntsToChg[i].style.fontSize) {
-				if (elmntsToChg[i].className.includes('title', 0)) {
-					// @ts-ignore
-					elmntsToChg[i].style.fontSize = "3.5rem";
-
-				} else if (elmntsToChg[i].className.includes('subtitle', 0)) {
-					// @ts-ignore
-					elmntsToChg[i].style.fontSize = "2.5rem";
-				
-				} else {
-					// @ts-ignore
-					elmntsToChg[i].style.fontSize = "1.5rem";
-					
-				}
-
-			} else {
-				// @ts-ignore
-				let temp = elmntsToChg[i].style.fontSize;
-				temp = parseFloat(temp);
-				temp = temp + 1;
-				let finalTemp = temp.toString() + 'rem';
-
-				// @ts-ignore
-				elmntsToChg[i].style.fontSize = finalTemp;
-
-
-			}
-
-		}
-
-	}
-	
-	function DecFontSize() {
-		var elmntsToChg = document.querySelectorAll('[id=text]');
-		
-		for (let i = 0; i < elmntsToChg.length; i ++) {
-			// @ts-ignore
-			if (!elmntsToChg[i].style.fontSize) {
-				if (elmntsToChg[i].className.includes('title', 0)) {
-					// @ts-ignore
-					elmntsToChg[i].style.fontSize = "3.5rem";
-
-				} else if (elmntsToChg[i].className.includes('subtitle', 0)) {
-					// @ts-ignore
-					elmntsToChg[i].style.fontSize = "2.5rem";
-				
-				} else {
-					// @ts-ignore
-					elmntsToChg[i].style.fontSize = "1.5rem";
-					
-				}
-
-			} else {
-				// @ts-ignore
-				let temp = elmntsToChg[i].style.fontSize;
-				temp = parseFloat(temp);
-				temp = temp - 1;
-				let finalTemp = temp.toString() + 'rem';
-
-				// @ts-ignore
-				elmntsToChg[i].style.fontSize = finalTemp;
-
-
-			}
-
-		}
-
-	}
-
 </script>
 
 <svelte:head>
 	<title>Log In</title>
 	<meta name="description" content="Log in to your account" />
 </svelte:head>
-
-<button class="PosButton" on:click={IncFontSize}>+</button>
-<button class="NegButton" on:click={DecFontSize}>-</button>
 
 <PageTransition>
   <div class="login-container">
@@ -108,9 +30,9 @@
 	  <p id="text">Enter your email and password</p>
 
 		<div class="form">
-			<label for="email">Email</label>
+			<label for="email" id="text">Email</label>
 			<input type="text" id="email" bind:value={email} />
-			<label for="password">Password</label>
+			<label for="password" id="text">Password</label>
 			<input type="password" id="password" bind:value={password} />
 
 			<div class="form-buttons">
@@ -175,17 +97,4 @@
 	outline: none;
 	border: 2px solid #000000;
   	}
-
-  	.PosButton {
-		position: absolute;
-		top: 15%;
-		left: 96%;
-	}
-
-	.NegButton {
-		position: absolute;
-		top: 15%;
-		left: 98%;
-	}
-
 </style>
